UploadWorker was relocated from app module to ui module
