+++ app/build.gradle
@@ -41,6 +41,7 @@ android {
 
 dependencies {
     implementation project(':domain')
+    implementation project(':data')
     implementation project(':ui')
 
     // Work Manager
