+++ app/src/main/AndroidManifest.xml
@@ -6,7 +6,7 @@
 
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
     <application
-        android:name=".workManager.WorkManagerApplication"
+        android:name=".WorkManagerApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
