+++ data/src/main/java/com/example/data/workManager/UploadWorker.kt
@@ -1,4 +1,4 @@
-package io.navendra.retrofitkotlindeferred.workManager
+package com.example.data.workManager
 
 import android.content.Context
 import androidx.work.*
